weaviate_utils/requirements.txt
pydantic
numpy
weaviate-client
classconfig
sqlalchemy
tqdm
psycopg2
aicaller~=1.1.6
